requirements.txt
# ══════════════════════════════════════════════════════════
# TRADING BOT - PYTHON DEPENDENCIES
# ══════════════════════════════════════════════════════════
# Install all: pip install -r requirements.txt
# ══════════════════════════════════════════════════════════


# ──────────────────────────────────────
# CORE
# ──────────────────────────────────────
python-dotenv==1.0.1
requests==2.31.0
schedule==1.2.1


# ──────────────────────────────────────
# DATABASE
# ──────────────────────────────────────
SQLAlchemy==2.0.25


# ──────────────────────────────────────
# DATA & ANALYSIS
# ──────────────────────────────────────
pandas==2.1.4
numpy==1.26.3
pandas-ta==0.3.14b


# ──────────────────────────────────────
# BROKER API
# ──────────────────────────────────────
dhanhq==2.0.1


# ──────────────────────────────────────
# TELEGRAM BOT
# ──────────────────────────────────────
python-telegram-bot==20.7


# ──────────────────────────────────────
# TIMEZONE
# ──────────────────────────────────────
pytz==2024.1


# ──────────────────────────────────────
# DEVELOPMENT & TESTING
# ──────────────────────────────────────
pytest==7.4.4